added more bullmq

+++ campa/src/SoundCronService.ts
@@ -6,11 +6,8 @@ import {
   type SoundCronJobEstablished,
   type SoundCronJob,
 } from "@discord-bigben/types";
-import { type WorkerRecordRepo } from "./WorkerRecordRepo";
 import type winston from "winston";
 import { type Redis } from "ioredis";
-import { debugLogger } from "./logging";
-import { UUID } from "crypto";
 
 type SoundCronQueue = bullmq.Queue<SoundCronJob, SoundCronJobEstablished>;
 
@@ -22,91 +19,13 @@ export class SoundCronService {
   private readonly soundCronQueue: SoundCronQueue;
 
   constructor(
-    private readonly workerMapRepo: WorkerRecordRepo,
     private readonly soundCronRepo: SoundCronRepo,
-    private readonly redis: Redis,
+    redis: Redis,
     private readonly logger: winston.Logger,
   ) {
     this.soundCronQueue = new bullmq.Queue("soundCron", {
-      connection: redis,
+      connection: redis
     });
-
-    /* We set up an event listener for when a soundcron job is sent. This
-    lets us know that we expect it to be alive, and we start listening for
-    heartbeats. */
-    new bullmq.QueueEvents("soundCron", {
-      connection: redis,
-    }).on("completed", async (job) => {
-      const retrievedJob = await bullmq.Job.fromId<
-        SoundCronJob,
-        SoundCronJobEstablished
-      >(this.soundCronQueue, job.jobId);
-      if (retrievedJob === undefined) {
-        this.logger.warn(
-          `Job ${job.jobId} completed, but could not be retrieved from the queue. This is unusual.`,
-        );
-        return;
-      }
-      const { workerId, key } = retrievedJob.returnvalue;
-      await this.workerMapRepo.addWorkerRecord(workerId, key);
-      const heartbeatInterval = 10000;
-      const timeout = setInterval(async () => {
-        if (await this.checkWorkerAlive(workerId)) {
-          debugLogger(`Worker ${workerId} is alive`);
-          return;
-        }
-        this.logger.warn(
-          `Worker ${workerId} has died. Attempting to resurrect soundcrons.`,
-        );
-        await redis.sadd("deadWorkers", workerId);
-        this.logger.warn(
-          `Worker ${workerId} has been marked as dead to prevent a zombie.`
-        );
-        await this.resurrectSoundCronsForWorker(workerId);
-        /* We assume that the resurrection works the first time. If it doesn't,
-        then the user will have to manually restart the soundcron. */
-        clearInterval(timeout);
-      }, heartbeatInterval);
-    });
-
-    setInterval(async () => {
-      await this.redis.setex("heartbeat:master", 10, 1);
-    }, 5000);
-  }
-
-  private async checkWorkerAlive(workerId: UUID): Promise<boolean> {
-    const exists = await this.redis.exists(`heartbeat:${workerId}`);
-    return exists === 1;
-  }
-
-  private async resurrectSoundCronsForWorker(workerId: UUID): Promise<void> {
-    const keys = await this.workerMapRepo.retrieveWorkerRecords(workerId);
-    await Promise.all(
-      keys.map(async (key) => {
-        // We want as many of these to succeed as possible, so we catch and log errors.
-        try {
-          const [serverId, name] = key.split(":", 2);
-          const soundCron = await this.soundCronRepo.getCron(serverId, name);
-          if (soundCron === null) {
-            const message = `Somehow there was a worker record for a soundcron that doesn't exist in the database: ${key}`;
-            this.logger.warn(message);
-            return;
-          }
-          await this.enqueueCron(serverId, soundCron);
-        } catch (err) {
-          if (err instanceof Error) {
-            this.logger.error(
-              `Error resurrecting soundcron ${key} for worker ${workerId}`,
-              err,
-            );
-          } else {
-            this.logger.error(
-              `Unknown error resurrecting soundcron ${key} for worker ${workerId} - error is not an instance of Error`,
-            );
-          }
-        }
-      }),
-    );
   }
 
   private async enqueueCron(
@@ -118,39 +37,24 @@ export class SoundCronService {
     } else if (soundCron.name === undefined) {
       throw new SoundCronServiceError("Soundcron name is undefined (not literally) - refusing to enqueue soundcron");
     }
+    const { name, timezone, cron } = soundCron;
     // This key is guaranteed to be unique by our database schema
-    const key = `${serverId}:${soundCron.name}`;
+    const jobName = `${serverId}:${soundCron.name}`;
     const jobData: SoundCronJob = {
       serverId,
-      name: soundCron.name,
-      cron: soundCron.cron,
-      timezone: soundCron.timezone ?? "UTC",
+      name,
+      cron,
+      timezone: timezone ?? "UTC",
       audio: soundCron.audio,
       mute: soundCron.mute ?? false,
       excludeChannels: soundCron.excludeChannels ?? [],
     };
-    await this.soundCronQueue.add(key, jobData);
-    await this.workerMapRepo.addUnassigned(key);
-  }
-
-  /**
-   * Starts all soundcrons. This will add all soundcrons to the queue.
-   * It does not write or modfiy any soundcrons in the database.
-   * It is recommended to only call this on startup. Every new soundcron
-   * will be added to the queue when it is added to the database with `addCron`.
-   */
-  async startAllCrons(): Promise<void> {
-    const allCrons = await this.soundCronRepo.listAllCrons();
-    await Promise.all(
-      Object.entries(allCrons).flatMap(([serverId, crons]) =>
-        crons.map(async (cron) => {
-          await this.enqueueCron(serverId, cron);
-        }),
-      ),
-    );
-    /* We do this explicitly to avoid a Promise<void[]>, which
-    linters dont like */
-    await Promise.resolve();
+    await this.soundCronQueue.add(jobName, jobData, {
+      repeat: {
+        pattern: cron,
+        tz: timezone
+      },
+    });
   }
 
   /**
@@ -208,26 +112,34 @@ export class SoundCronService {
    * @throws SoundCronServiceError if there is an error removing the soundcron from the database or queue
    */
   async removeCron(serverId: string, name: string): Promise<void> {
-    try {
-      await this.soundCronRepo.removeCron(serverId, name);
-    } catch (err) {
-      if (err instanceof Error) {
-        throw new SoundCronServiceError(
-          `Unable to remove from database`,
-          err,
-        );
-      }
+    /* We do make 2 round trips to the database here, even though
+    MariaDB makes it easy to do it in one. This is fine because
+    removal is not a common operation */
+    const soundCron = await this.soundCronRepo.getCron(serverId, name);
+
+    /* In the future, we should probably return something to the caller
+    indicating that this happened. It isn't an application error though,
+    so an exception is not appropriate */
+    if (soundCron === null) {
+      this.logger.info(`Attempt to remove non-existent soundcron ${name} for server ${serverId}`);
+      return;
     }
 
-    const key = `${serverId}:${name}`;
-    await this.redis.sadd("removedSoundCrons", key);
+    const jobName = `${serverId}:${name}`; 
+    const removedFromQueue = this.soundCronQueue.removeRepeatable(jobName, {
+      pattern: soundCron.cron,
+      tz: soundCron.timezone,
+    });
+
+    const removedFromDatabase = this.soundCronRepo.removeCron(serverId, name);
+  
+    /* We run these two in tandem for two reasons:
+    1. Performance (minor)
+    2. Even if one fails, we would hope the other would succeed */
+    await Promise.all([removedFromQueue, removedFromDatabase]);
   }
 
   async listCrons(serverId: string): Promise<SoundCron[]> {
     return await this.soundCronRepo.listCrons(serverId);
-  }
-
-  async listAllCrons(): Promise<Record<string, SoundCron[]>> {
-    return await this.soundCronRepo.listAllCrons();
   }
 }
